#85 add markdown rendering support for dataset and model views, create views folder in pkpdapps
