+++ pkpdapp/pkpdapp/templates/dataset_detail.html
@@ -6,10 +6,34 @@ copyright notice and full license details.
 
 {% extends "base/base.html" %}
 {% load static %}
+{% load markdownify %}
 
 {% block main_content %}
 <p></p>
 <h2>Dataset {{dataset.name}}</h2>
-<p>{{dataset.description}}</p>
+<p>{{dataset.description|markdownify}}</p>
+<h3>Biomarker Types</h2>
+<ul>
+{% for type in biomarker_types %}
+  <li>{{ type.name }}: {{ type.description|truncatewords:"10"}}</li>
+{% endfor %}
+</ul>
+<h3>Data</h2>
+<table style="width:100%">
+  <tr>
+    <th>Time</th>
+    <th>Subject Id</th>
+    <th>Biomarker Type</th>
+    <th>Value</th>
+  </tr>
+  {% for biomarker in biomarker_dataset %}
+  <tr>
+    <td>{{ biomarker.time }}</td>
+    <td>{{ biomarker.subject_id}}</td>
+    <td>{{ biomarker.biomarker_type.name }}</td>
+    <td>{{ biomarker.value }}</td>
+  </tr>
+  {% endfor %}
+</table>
 <p></p>
 {% endblock %}
